cloud-messaging-project/requirements.txt
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
flask-cors==6.0.2
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-pubsub==2.35.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
requests==2.32.5
rsa==4.9.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
zipp==3.23.0